fix: исправить расширение файлов и использование кнопки
