+++ src/components/Button/Button.jsx
@@ -1,25 +1,26 @@
 import PropTypes from 'prop-types';
 import './Button.scss';
 
-export const Button = ({ primary, border, size, label, iconType, ...props }) => {
+export const Button = ({ primary, border, rotate, size, type, children, ...props }) => {
   const mode = primary ? 'button__primary' : 'button__secondary';
   const isBorder = border ? `${mode}_border` : '';
-  // (size === 'l') ?
+  const isRotate = rotate && size === 'xs' ? 'button_rotate' : '';
   return (
     <button
-      type="button"
-      className={['button', `button_size_${size}`, mode, isBorder].join(' ')}
+      type={type}
+      className={['button', `button_size_${size}`, mode, isBorder, isRotate].join(' ')}
       {...props}>
-      {label}
+      {children}
     </button>
   );
 };
 
 Button.propTypes = {
   primary: PropTypes.bool,
-  size: PropTypes.oneOf(['s', 'm', 'l', 'xl', 'xxl']),
-  label: PropTypes.string.isRequired,
+  size: PropTypes.oneOf(['xs', 's', 'm', 'l', 'xl', 'xxl', 'xxx', 'xxxxl']),
   border: PropTypes.bool,
+  rotate: PropTypes.bool,
+  type: PropTypes.oneOf(['button', 'submit']),
   onClick: PropTypes.func,
 };
 
@@ -27,5 +28,6 @@ Button.defaultProps = {
   primary: false,
   size: 'm',
   border: true,
+  rotate: false,
   onClick: undefined,
 };
